+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/bean/dto/BankCardUpdateInfo.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.familyhelper.finance.stack.bean.dto;
 
 import com.dwarfeng.subgrade.stack.bean.dto.Dto;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 
 /**
  * 银行卡创更新信息。
@@ -10,8 +11,9 @@ import com.dwarfeng.subgrade.stack.bean.dto.Dto;
  */
 public class BankCardUpdateInfo implements Dto {
 
-    private static final long serialVersionUID = 5331634805509187489L;
+    private static final long serialVersionUID = -5253559119807476705L;
 
+    private LongIdKey bankCardKey;
     private String name;
     private String cardType;
     private String remark;
@@ -19,10 +21,19 @@ public class BankCardUpdateInfo implements Dto {
     public BankCardUpdateInfo() {
     }
 
-    public BankCardUpdateInfo(String name, String cardType, String remark) {
+    public BankCardUpdateInfo(LongIdKey bankCardKey, String name, String cardType, String remark) {
+        this.bankCardKey = bankCardKey;
         this.name = name;
         this.cardType = cardType;
         this.remark = remark;
+    }
+
+    public LongIdKey getBankCardKey() {
+        return bankCardKey;
+    }
+
+    public void setBankCardKey(LongIdKey bankCardKey) {
+        this.bankCardKey = bankCardKey;
     }
 
     public String getName() {
@@ -52,7 +63,8 @@ public class BankCardUpdateInfo implements Dto {
     @Override
     public String toString() {
         return "BankCardUpdateInfo{" +
-                "name='" + name + '\'' +
+                "bankCardKey=" + bankCardKey +
+                ", name='" + name + '\'' +
                 ", cardType='" + cardType + '\'' +
                 ", remark='" + remark + '\'' +
                 '}';