+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/handler/BankCardOperateHandler.java
@@ -19,24 +19,20 @@ public interface BankCardOperateHandler extends Handler {
      * 创建银行卡。
      *
      * @param userKey            银行卡的所有者的主键。
-     * @param accountBookKey     银行卡所属的账户的主键。
      * @param bankCardCreateInfo 银行卡的创建信息。
      * @return 生成的银行卡的主键。
      * @throws HandlerException 处理器异常。
      */
-    LongIdKey createBankCard(StringIdKey userKey, LongIdKey accountBookKey, BankCardCreateInfo bankCardCreateInfo)
-            throws HandlerException;
+    LongIdKey createBankCard(StringIdKey userKey, BankCardCreateInfo bankCardCreateInfo) throws HandlerException;
 
     /**
      * 更新银行卡。
      *
      * @param userKey            银行卡的所有者的主键。
-     * @param bankCardKey        银行卡的主键。
      * @param bankCardUpdateInfo 银行卡的更新信息。
      * @throws HandlerException 处理器异常。
      */
-    void updateBankCard(StringIdKey userKey, LongIdKey bankCardKey, BankCardUpdateInfo bankCardUpdateInfo)
-            throws HandlerException;
+    void updateBankCard(StringIdKey userKey, BankCardUpdateInfo bankCardUpdateInfo) throws HandlerException;
 
     /**
      * 创建银行卡。